Module Pbrt_quickcheck

Runtime for QuickCheck based tests.

This module contains additional types and functions that are used by the code generator when the option --quickcheck is supplied to ocaml-protoc.

module Type_class : sig ... end

A type class generated for each type

module Test : sig ... end