Module Pbrt_quickcheck.Test

type t = QCheck2.Test.t

val make : ?gen:'a QCheck2.Gen.t -> 'a Type_class.t -> t list

Generates a test suite for that type that checks that values roundtrip through serializations. The generator to use can be overriden by supplying gen.